Set bg color of main depending on child page
This is more of a workaround to set the background color for the question pages so that the blue background expands to the footer. Expanding the whole page <div> is not easily possible with CSS since the <main> element is child of a flex and does not have a pre-defined height. If one finds a better solution, please feel free to replace this.